handle update filters
